New Feature: Custom functions on hitting both Y or Z endstops =============================================================
This FAB-UI functionality is linked to a still unmerged FABlin pull-request and will not work without it.

It provides flexibility to interact with FABUI directly from the FABtotum, without the need of a client browser.

Currently only "shutdown" is implemented. For it to work you have to select it under settings and restart your FABtotum.

The Settings implementation allows for selection of the action to be undertaken. Currently only shutdown is implemented.